Added button to refresh the list on the main page and added progress checking for the AppStateThread
Added a button on the home page to refresh the list of items that
displays Venues as a placeholder for now. Also changed the
AppStateThread variable to be static now instead of the AppState inside
of the Thread. Can also use AppStateThread.getStatus() method to check
if the process is complete.